+++ crates/iceberg/src/arrow/reader.rs
@@ -27,8 +27,7 @@ use arrow_array::{Array, ArrayRef, BooleanArray, Datum as ArrowDatum, RecordBatc
 use arrow_cast::cast::cast;
 use arrow_ord::cmp::{eq, gt, gt_eq, lt, lt_eq, neq};
 use arrow_schema::{
-    ArrowError, DataType, Field, FieldRef, Fields, Schema as ArrowSchema,
-    SchemaRef as ArrowSchemaRef, TimeUnit,
+    ArrowError, DataType, FieldRef, Schema as ArrowSchema, SchemaRef as ArrowSchemaRef,
 };
 use arrow_string::like::starts_with;
 use bytes::Bytes;
@@ -47,8 +46,8 @@ use parquet::schema::types::{SchemaDescriptor, Type as ParquetType};
 use typed_builder::TypedBuilder;
 
 use crate::arrow::caching_delete_file_loader::CachingDeleteFileLoader;
+use crate::arrow::int96::coerce_int96_timestamps;
 use crate::arrow::record_batch_transformer::RecordBatchTransformerBuilder;
-use crate::arrow::schema::{ArrowSchemaVisitor, DEFAULT_MAP_FIELD_NAME, visit_schema};
 use crate::arrow::{arrow_schema_to_schema, get_arrow_datum};
 use crate::delete_vector::DeleteVector;
 use crate::error::Result;
@@ -1255,203 +1254,6 @@ fn add_fallback_field_ids_to_arrow_schema(arrow_schema: &ArrowSchemaRef) -> Arc<
     ))
 }
 
-/// Coerce Arrow schema types for INT96 columns to match the Iceberg table schema.
-///
-/// arrow-rs defaults INT96 to `Timestamp(Nanosecond)`, which overflows i64 for dates outside
-/// ~1677-2262. We use arrow-rs's schema hint mechanism to read INT96 at the resolution
-/// specified by the Iceberg schema (`timestamp` → microsecond, `timestamp_ns` → nanosecond).
-///
-/// Iceberg Java handles this differently: it bypasses parquet-mr with a custom column reader
-/// (`GenericParquetReaders.TimestampInt96Reader`). We achieve the same result via schema hints.
-///
-/// References:
-/// - Iceberg spec primitive types: <https://iceberg.apache.org/spec/#primitive-types>
-/// - arrow-rs schema hint support: <https://github.com/apache/arrow-rs/pull/7285>
-fn coerce_int96_timestamps(
-    arrow_schema: &ArrowSchemaRef,
-    iceberg_schema: &Schema,
-) -> Option<Arc<ArrowSchema>> {
-    let mut visitor = Int96CoercionVisitor::new(iceberg_schema);
-    let coerced = visit_schema(arrow_schema, &mut visitor).ok()?;
-    if visitor.changed {
-        Some(Arc::new(coerced))
-    } else {
-        None
-    }
-}
-
-/// Visitor that coerces `Timestamp(Nanosecond)` Arrow fields to the resolution
-/// indicated by the Iceberg schema. Follows the same pattern as `MetadataStripVisitor`.
-struct Int96CoercionVisitor<'a> {
-    iceberg_schema: &'a Schema,
-    field_stack: Vec<Field>,
-    changed: bool,
-}
-
-impl<'a> Int96CoercionVisitor<'a> {
-    fn new(iceberg_schema: &'a Schema) -> Self {
-        Self {
-            iceberg_schema,
-            field_stack: Vec::new(),
-            changed: false,
-        }
-    }
-
-    /// Determine the target TimeUnit for a Timestamp(Nanosecond) field based on the
-    /// Iceberg schema. Falls back to microsecond when field IDs are unavailable,
-    /// matching Iceberg Java behavior.
-    fn target_unit(&self, field: &Field) -> Option<TimeUnit> {
-        if !matches!(
-            field.data_type(),
-            DataType::Timestamp(TimeUnit::Nanosecond, _)
-        ) {
-            return None;
-        }
-
-        let target = field
-            .metadata()
-            .get(PARQUET_FIELD_ID_META_KEY)
-            .and_then(|id_str| id_str.parse::<i32>().ok())
-            .and_then(|field_id| self.iceberg_schema.field_by_id(field_id))
-            .and_then(|f| match &*f.field_type {
-                Type::Primitive(PrimitiveType::Timestamp | PrimitiveType::Timestamptz) => {
-                    Some(TimeUnit::Microsecond)
-                }
-                Type::Primitive(PrimitiveType::TimestampNs | PrimitiveType::TimestamptzNs) => {
-                    Some(TimeUnit::Nanosecond)
-                }
-                _ => None,
-            })
-            // Iceberg Java reads INT96 as microseconds by default
-            .unwrap_or(TimeUnit::Microsecond);
-
-        if target == TimeUnit::Nanosecond {
-            None
-        } else {
-            Some(target)
-        }
-    }
-}
-
-impl ArrowSchemaVisitor for Int96CoercionVisitor<'_> {
-    type T = Field;
-    type U = ArrowSchema;
-
-    fn before_field(&mut self, field: &Field) -> Result<()> {
-        self.field_stack.push(field.as_ref().clone());
-        Ok(())
-    }
-
-    fn before_list_element(&mut self, field: &Field) -> Result<()> {
-        self.field_stack.push(field.as_ref().clone());
-        Ok(())
-    }
-
-    fn before_map_key(&mut self, field: &Field) -> Result<()> {
-        self.field_stack.push(field.as_ref().clone());
-        Ok(())
-    }
-
-    fn before_map_value(&mut self, field: &Field) -> Result<()> {
-        self.field_stack.push(field.as_ref().clone());
-        Ok(())
-    }
-
-    fn schema(&mut self, schema: &ArrowSchema, values: Vec<Self::T>) -> Result<Self::U> {
-        Ok(ArrowSchema::new_with_metadata(
-            values,
-            schema.metadata().clone(),
-        ))
-    }
-
-    fn r#struct(&mut self, _fields: &Fields, results: Vec<Self::T>) -> Result<Self::T> {
-        let field_info = self
-            .field_stack
-            .pop()
-            .ok_or_else(|| Error::new(ErrorKind::Unexpected, "Field stack underflow in struct"))?;
-        Ok(Field::new(
-            field_info.name(),
-            DataType::Struct(Fields::from(results)),
-            field_info.is_nullable(),
-        )
-        .with_metadata(field_info.metadata().clone()))
-    }
-
-    fn list(&mut self, list: &DataType, value: Self::T) -> Result<Self::T> {
-        let field_info = self
-            .field_stack
-            .pop()
-            .ok_or_else(|| Error::new(ErrorKind::Unexpected, "Field stack underflow in list"))?;
-        let list_type = match list {
-            DataType::List(_) => DataType::List(Arc::new(value)),
-            DataType::LargeList(_) => DataType::LargeList(Arc::new(value)),
-            DataType::FixedSizeList(_, size) => DataType::FixedSizeList(Arc::new(value), *size),
-            _ => {
-                return Err(Error::new(
-                    ErrorKind::Unexpected,
-                    format!("Expected list type, got {list}"),
-                ));
-            }
-        };
-        Ok(
-            Field::new(field_info.name(), list_type, field_info.is_nullable())
-                .with_metadata(field_info.metadata().clone()),
-        )
-    }
-
-    fn map(&mut self, map: &DataType, key_value: Self::T, value: Self::T) -> Result<Self::T> {
-        let field_info = self
-            .field_stack
-            .pop()
-            .ok_or_else(|| Error::new(ErrorKind::Unexpected, "Field stack underflow in map"))?;
-        let sorted = match map {
-            DataType::Map(_, sorted) => *sorted,
-            _ => {
-                return Err(Error::new(
-                    ErrorKind::Unexpected,
-                    format!("Expected map type, got {map}"),
-                ));
-            }
-        };
-        let struct_field = Field::new(
-            DEFAULT_MAP_FIELD_NAME,
-            DataType::Struct(Fields::from(vec![key_value, value])),
-            false,
-        );
-        Ok(Field::new(
-            field_info.name(),
-            DataType::Map(Arc::new(struct_field), sorted),
-            field_info.is_nullable(),
-        )
-        .with_metadata(field_info.metadata().clone()))
-    }
-
-    fn primitive(&mut self, p: &DataType) -> Result<Self::T> {
-        let field_info = self.field_stack.pop().ok_or_else(|| {
-            Error::new(ErrorKind::Unexpected, "Field stack underflow in primitive")
-        })?;
-
-        if let Some(target_unit) = self.target_unit(&field_info) {
-            let tz = match field_info.data_type() {
-                DataType::Timestamp(_, tz) => tz.clone(),
-                _ => None,
-            };
-            self.changed = true;
-            Ok(Field::new(
-                field_info.name(),
-                DataType::Timestamp(target_unit, tz),
-                field_info.is_nullable(),
-            )
-            .with_metadata(field_info.metadata().clone()))
-        } else {
-            Ok(
-                Field::new(field_info.name(), p.clone(), field_info.is_nullable())
-                    .with_metadata(field_info.metadata().clone()),
-            )
-        }
-    }
-}
-
 /// A visitor to collect field ids from bound predicates.
 struct CollectFieldIdVisitor {
     field_ids: HashSet<i32>,
@@ -2178,20 +1980,16 @@ mod tests {
     use std::sync::Arc;
 
     use arrow_array::cast::AsArray;
-    use arrow_array::{
-        ArrayRef, LargeStringArray, RecordBatch, StringArray, TimestampMicrosecondArray,
-    };
+    use arrow_array::{ArrayRef, LargeStringArray, RecordBatch, StringArray};
     use arrow_schema::{DataType, Field, Schema as ArrowSchema, TimeUnit};
     use futures::TryStreamExt;
     use parquet::arrow::arrow_reader::{RowSelection, RowSelector};
     use parquet::arrow::{ArrowWriter, ProjectionMask};
-    use parquet::basic::{Compression, Repetition, Type as PhysicalType};
-    use parquet::data_type::{ByteArrayType, Int32Type, Int96, Int96Type};
+    use parquet::basic::Compression;
     use parquet::file::metadata::{ColumnChunkMetaData, RowGroupMetaData};
     use parquet::file::properties::WriterProperties;
-    use parquet::file::writer::SerializedFileWriter;
     use parquet::schema::parser::parse_message_type;
-    use parquet::schema::types::{SchemaDescPtr, SchemaDescriptor, Type as SchemaType};
+    use parquet::schema::types::{SchemaDescPtr, SchemaDescriptor};
     use roaring::RoaringTreemap;
     use tempfile::TempDir;
 
@@ -4890,542 +4688,5 @@ message schema {
         assert_eq!(result[0], expected_0);
         assert_eq!(result[1], expected_1);
         assert_eq!(result[2], expected_2);
-    }
-
-    // INT96 encoding: [nanos_low_u32, nanos_high_u32, julian_day_u32]
-    // Julian day 2_440_588 = Unix epoch (1970-01-01)
-    const UNIX_EPOCH_JULIAN: i64 = 2_440_588;
-    const MICROS_PER_DAY: i64 = 86_400_000_000;
-    /// Noon on 3333-01-01 (Julian day 2_953_529) — outside the i64 nanosecond range (~1677-2262).
-    const INT96_TEST_NANOS_WITHIN_DAY: u64 = 43_200_000_000_000;
-    const INT96_TEST_JULIAN_DAY: u32 = 2_953_529;
-
-    /// Build an INT96 value and its expected microsecond interpretation.
-    fn make_int96_test_value() -> (Int96, i64) {
-        let mut val = Int96::new();
-        val.set_data(
-            (INT96_TEST_NANOS_WITHIN_DAY & 0xFFFFFFFF) as u32,
-            (INT96_TEST_NANOS_WITHIN_DAY >> 32) as u32,
-            INT96_TEST_JULIAN_DAY,
-        );
-        let expected_micros = (INT96_TEST_JULIAN_DAY as i64 - UNIX_EPOCH_JULIAN) * MICROS_PER_DAY
-            + (INT96_TEST_NANOS_WITHIN_DAY / 1_000) as i64;
-        (val, expected_micros)
-    }
-
-    /// Read a Parquet file through ArrowReader and return the resulting batches.
-    async fn read_int96_batches(
-        file_path: &str,
-        schema: SchemaRef,
-        project_field_ids: Vec<i32>,
-    ) -> Vec<RecordBatch> {
-        let file_io = FileIO::new_with_fs();
-        let reader = ArrowReaderBuilder::new(file_io).build();
-
-        let file_size = std::fs::metadata(file_path).unwrap().len();
-        let task = FileScanTask {
-            file_size_in_bytes: file_size,
-            start: 0,
-            length: file_size,
-            record_count: None,
-            data_file_path: file_path.to_string(),
-            data_file_format: DataFileFormat::Parquet,
-            schema,
-            project_field_ids,
-            predicate: None,
-            deletes: vec![],
-            partition: None,
-            partition_spec: None,
-            name_mapping: None,
-            case_sensitive: false,
-        };
-
-        let tasks = Box::pin(futures::stream::iter(vec![Ok(task)])) as FileScanTaskStream;
-        reader.read(tasks).unwrap().try_collect().await.unwrap()
-    }
-
-    /// Writes a Parquet file with INT96 timestamps using SerializedFileWriter
-    /// (ArrowWriter cannot write INT96). Returns (file_path, expected_microsecond_values).
-    fn write_int96_parquet_file(
-        table_location: &str,
-        filename: &str,
-        with_field_ids: bool,
-    ) -> (String, Vec<i64>) {
-        let file_path = format!("{table_location}/{filename}");
-
-        let mut ts_builder = SchemaType::primitive_type_builder("ts", PhysicalType::INT96)
-            .with_repetition(Repetition::OPTIONAL);
-        let mut id_builder = SchemaType::primitive_type_builder("id", PhysicalType::INT32)
-            .with_repetition(Repetition::REQUIRED);
-
-        if with_field_ids {
-            ts_builder = ts_builder.with_id(Some(1));
-            id_builder = id_builder.with_id(Some(2));
-        }
-
-        let schema = SchemaType::group_type_builder("schema")
-            .with_fields(vec![
-                Arc::new(ts_builder.build().unwrap()),
-                Arc::new(id_builder.build().unwrap()),
-            ])
-            .build()
-            .unwrap();
-
-        // Dates outside the i64 nanosecond range (~1677-2262) overflow without coercion.
-        const NOON_NANOS: u64 = INT96_TEST_NANOS_WITHIN_DAY;
-        const JULIAN_3333: u32 = INT96_TEST_JULIAN_DAY;
-        const JULIAN_2100: u32 = 2_488_070;
-
-        let test_data: Vec<(u32, u32, u32, i64)> = vec![
-            // 3333-01-01 00:00:00
-            (
-                0,
-                0,
-                JULIAN_3333,
-                (JULIAN_3333 as i64 - UNIX_EPOCH_JULIAN) * MICROS_PER_DAY,
-            ),
-            // 3333-01-01 12:00:00
-            (
-                (NOON_NANOS & 0xFFFFFFFF) as u32,
-                (NOON_NANOS >> 32) as u32,
-                JULIAN_3333,
-                (JULIAN_3333 as i64 - UNIX_EPOCH_JULIAN) * MICROS_PER_DAY
-                    + (NOON_NANOS / 1_000) as i64,
-            ),
-            // 2100-01-01 00:00:00
-            (
-                0,
-                0,
-                JULIAN_2100,
-                (JULIAN_2100 as i64 - UNIX_EPOCH_JULIAN) * MICROS_PER_DAY,
-            ),
-        ];
-
-        let int96_values: Vec<Int96> = test_data
-            .iter()
-            .map(|(lo, hi, day, _)| {
-                let mut v = Int96::new();
-                v.set_data(*lo, *hi, *day);
-                v
-            })
-            .collect();
-
-        let id_values: Vec<i32> = (0..test_data.len() as i32).collect();
-        let expected_micros: Vec<i64> = test_data.iter().map(|(_, _, _, m)| *m).collect();
-
-        let file = File::create(&file_path).unwrap();
-        let mut writer =
-            SerializedFileWriter::new(file, Arc::new(schema), Default::default()).unwrap();
-
-        let mut row_group = writer.next_row_group().unwrap();
-        {
-            // def=1: ts is OPTIONAL and present. No repetition levels (top-level columns).
-            let mut col = row_group.next_column().unwrap().unwrap();
-            col.typed::<Int96Type>()
-                .write_batch(&int96_values, Some(&vec![1; test_data.len()]), None)
-                .unwrap();
-            col.close().unwrap();
-        }
-        {
-            let mut col = row_group.next_column().unwrap().unwrap();
-            col.typed::<Int32Type>()
-                .write_batch(&id_values, None, None)
-                .unwrap();
-            col.close().unwrap();
-        }
-        row_group.close().unwrap();
-        writer.close().unwrap();
-
-        (file_path, expected_micros)
-    }
-
-    /// Read INT96 Parquet file through ArrowReader and verify top-level microsecond timestamps.
-    async fn assert_int96_read_matches(
-        file_path: &str,
-        schema: SchemaRef,
-        project_field_ids: Vec<i32>,
-        expected_micros: &[i64],
-    ) {
-        let batches = read_int96_batches(file_path, schema, project_field_ids).await;
-
-        assert_eq!(batches.len(), 1);
-        let ts_array = batches[0]
-            .column(0)
-            .as_any()
-            .downcast_ref::<TimestampMicrosecondArray>()
-            .expect("Expected TimestampMicrosecondArray");
-
-        for (i, expected) in expected_micros.iter().enumerate() {
-            assert_eq!(
-                ts_array.value(i),
-                *expected,
-                "Row {i}: got {}, expected {expected}",
-                ts_array.value(i)
-            );
-        }
-    }
-
-    /// Files with embedded field IDs (branch 1 of schema resolution).
-    #[tokio::test]
-    async fn test_read_int96_timestamps_with_field_ids() {
-        let schema = Arc::new(
-            Schema::builder()
-                .with_schema_id(1)
-                .with_fields(vec![
-                    NestedField::optional(1, "ts", Type::Primitive(PrimitiveType::Timestamp))
-                        .into(),
-                    NestedField::required(2, "id", Type::Primitive(PrimitiveType::Int)).into(),
-                ])
-                .build()
-                .unwrap(),
-        );
-
-        let tmp_dir = TempDir::new().unwrap();
-        let table_location = tmp_dir.path().to_str().unwrap().to_string();
-        let (file_path, expected_micros) =
-            write_int96_parquet_file(&table_location, "with_ids.parquet", true);
-
-        assert_int96_read_matches(&file_path, schema, vec![1, 2], &expected_micros).await;
-    }
-
-    /// Migrated files without field IDs (branches 2/3 of schema resolution).
-    #[tokio::test]
-    async fn test_read_int96_timestamps_without_field_ids() {
-        let schema = Arc::new(
-            Schema::builder()
-                .with_schema_id(1)
-                .with_fields(vec![
-                    NestedField::optional(1, "ts", Type::Primitive(PrimitiveType::Timestamp))
-                        .into(),
-                    NestedField::required(2, "id", Type::Primitive(PrimitiveType::Int)).into(),
-                ])
-                .build()
-                .unwrap(),
-        );
-
-        let tmp_dir = TempDir::new().unwrap();
-        let table_location = tmp_dir.path().to_str().unwrap().to_string();
-        let (file_path, expected_micros) =
-            write_int96_parquet_file(&table_location, "no_ids.parquet", false);
-
-        assert_int96_read_matches(&file_path, schema, vec![1, 2], &expected_micros).await;
-    }
-
-    /// Test reading INT96 timestamps inside a struct field.
-    #[tokio::test]
-    async fn test_read_int96_timestamps_in_struct() {
-        let tmp_dir = TempDir::new().unwrap();
-        let table_location = tmp_dir.path().to_str().unwrap().to_string();
-        let file_path = format!("{table_location}/struct_int96.parquet");
-
-        let ts_type = SchemaType::primitive_type_builder("ts", PhysicalType::INT96)
-            .with_repetition(Repetition::OPTIONAL)
-            .with_id(Some(2))
-            .build()
-            .unwrap();
-
-        let struct_type = SchemaType::group_type_builder("data")
-            .with_repetition(Repetition::REQUIRED)
-            .with_id(Some(1))
-            .with_fields(vec![Arc::new(ts_type)])
-            .build()
-            .unwrap();
-
-        let parquet_schema = SchemaType::group_type_builder("schema")
-            .with_fields(vec![Arc::new(struct_type)])
-            .build()
-            .unwrap();
-
-        let (int96_val, expected_micros) = make_int96_test_value();
-
-        let file = File::create(&file_path).unwrap();
-        let mut writer =
-            SerializedFileWriter::new(file, Arc::new(parquet_schema), Default::default()).unwrap();
-
-        // def=1: struct is REQUIRED so no level, ts is OPTIONAL and present (1).
-        // No repetition levels needed (no repeated groups).
-        let mut row_group = writer.next_row_group().unwrap();
-        {
-            let mut col = row_group.next_column().unwrap().unwrap();
-            col.typed::<Int96Type>()
-                .write_batch(&[int96_val], Some(&[1]), None)
-                .unwrap();
-            col.close().unwrap();
-        }
-        row_group.close().unwrap();
-        writer.close().unwrap();
-
-        let iceberg_schema = Arc::new(
-            Schema::builder()
-                .with_schema_id(1)
-                .with_fields(vec![
-                    NestedField::required(
-                        1,
-                        "data",
-                        Type::Struct(crate::spec::StructType::new(vec![
-                            NestedField::optional(
-                                2,
-                                "ts",
-                                Type::Primitive(PrimitiveType::Timestamp),
-                            )
-                            .into(),
-                        ])),
-                    )
-                    .into(),
-                ])
-                .build()
-                .unwrap(),
-        );
-
-        let batches = read_int96_batches(&file_path, iceberg_schema, vec![1]).await;
-
-        assert_eq!(batches.len(), 1);
-        let struct_array = batches[0]
-            .column(0)
-            .as_any()
-            .downcast_ref::<arrow_array::StructArray>()
-            .expect("Expected StructArray");
-        let ts_array = struct_array
-            .column(0)
-            .as_any()
-            .downcast_ref::<TimestampMicrosecondArray>()
-            .expect("Expected TimestampMicrosecondArray inside struct");
-
-        assert_eq!(
-            ts_array.value(0),
-            expected_micros,
-            "INT96 in struct: got {}, expected {expected_micros}",
-            ts_array.value(0)
-        );
-    }
-
-    /// Test reading INT96 timestamps inside a list field (3-level Parquet LIST encoding).
-    #[tokio::test]
-    async fn test_read_int96_timestamps_in_list() {
-        let tmp_dir = TempDir::new().unwrap();
-        let table_location = tmp_dir.path().to_str().unwrap().to_string();
-        let file_path = format!("{table_location}/list_int96.parquet");
-
-        // 3-level LIST encoding:
-        //   optional group timestamps (LIST) {
-        //     repeated group list {
-        //       optional int96 element;
-        //     }
-        //   }
-        let element_type = SchemaType::primitive_type_builder("element", PhysicalType::INT96)
-            .with_repetition(Repetition::OPTIONAL)
-            .with_id(Some(2))
-            .build()
-            .unwrap();
-
-        let list_group = SchemaType::group_type_builder("list")
-            .with_repetition(Repetition::REPEATED)
-            .with_fields(vec![Arc::new(element_type)])
-            .build()
-            .unwrap();
-
-        let list_type = SchemaType::group_type_builder("timestamps")
-            .with_repetition(Repetition::OPTIONAL)
-            .with_id(Some(1))
-            .with_logical_type(Some(parquet::basic::LogicalType::List))
-            .with_fields(vec![Arc::new(list_group)])
-            .build()
-            .unwrap();
-
-        let parquet_schema = SchemaType::group_type_builder("schema")
-            .with_fields(vec![Arc::new(list_type)])
-            .build()
-            .unwrap();
-
-        let (int96_val, expected_micros) = make_int96_test_value();
-
-        let file = File::create(&file_path).unwrap();
-        let mut writer =
-            SerializedFileWriter::new(file, Arc::new(parquet_schema), Default::default()).unwrap();
-
-        // Write a single row with a list containing one INT96 element.
-        // def=3: list present (1) + repeated group (2) + element present (3)
-        // rep=0: start of a new list
-        let mut row_group = writer.next_row_group().unwrap();
-        {
-            let mut col = row_group.next_column().unwrap().unwrap();
-            col.typed::<Int96Type>()
-                .write_batch(&[int96_val], Some(&[3]), Some(&[0]))
-                .unwrap();
-            col.close().unwrap();
-        }
-        row_group.close().unwrap();
-        writer.close().unwrap();
-
-        let iceberg_schema = Arc::new(
-            Schema::builder()
-                .with_schema_id(1)
-                .with_fields(vec![
-                    NestedField::optional(
-                        1,
-                        "timestamps",
-                        Type::List(crate::spec::ListType {
-                            element_field: NestedField::optional(
-                                2,
-                                "element",
-                                Type::Primitive(PrimitiveType::Timestamp),
-                            )
-                            .into(),
-                        }),
-                    )
-                    .into(),
-                ])
-                .build()
-                .unwrap(),
-        );
-
-        let batches = read_int96_batches(&file_path, iceberg_schema, vec![1]).await;
-
-        assert_eq!(batches.len(), 1);
-        let list_array = batches[0]
-            .column(0)
-            .as_any()
-            .downcast_ref::<arrow_array::ListArray>()
-            .expect("Expected ListArray");
-        let ts_array = list_array
-            .values()
-            .as_any()
-            .downcast_ref::<TimestampMicrosecondArray>()
-            .expect("Expected TimestampMicrosecondArray inside list");
-
-        assert_eq!(
-            ts_array.value(0),
-            expected_micros,
-            "INT96 in list: got {}, expected {expected_micros}",
-            ts_array.value(0)
-        );
-    }
-
-    /// Test reading INT96 timestamps as map values.
-    #[tokio::test]
-    async fn test_read_int96_timestamps_in_map() {
-        let tmp_dir = TempDir::new().unwrap();
-        let table_location = tmp_dir.path().to_str().unwrap().to_string();
-        let file_path = format!("{table_location}/map_int96.parquet");
-
-        // MAP encoding:
-        //   optional group ts_map (MAP) {
-        //     repeated group key_value {
-        //       required binary key (UTF8);
-        //       optional int96 value;
-        //     }
-        //   }
-        let key_type = SchemaType::primitive_type_builder("key", PhysicalType::BYTE_ARRAY)
-            .with_repetition(Repetition::REQUIRED)
-            .with_logical_type(Some(parquet::basic::LogicalType::String))
-            .with_id(Some(2))
-            .build()
-            .unwrap();
-
-        let value_type = SchemaType::primitive_type_builder("value", PhysicalType::INT96)
-            .with_repetition(Repetition::OPTIONAL)
-            .with_id(Some(3))
-            .build()
-            .unwrap();
-
-        let key_value_group = SchemaType::group_type_builder("key_value")
-            .with_repetition(Repetition::REPEATED)
-            .with_fields(vec![Arc::new(key_type), Arc::new(value_type)])
-            .build()
-            .unwrap();
-
-        let map_type = SchemaType::group_type_builder("ts_map")
-            .with_repetition(Repetition::OPTIONAL)
-            .with_id(Some(1))
-            .with_logical_type(Some(parquet::basic::LogicalType::Map))
-            .with_fields(vec![Arc::new(key_value_group)])
-            .build()
-            .unwrap();
-
-        let parquet_schema = SchemaType::group_type_builder("schema")
-            .with_fields(vec![Arc::new(map_type)])
-            .build()
-            .unwrap();
-
-        let (int96_val, expected_micros) = make_int96_test_value();
-
-        let file = File::create(&file_path).unwrap();
-        let mut writer =
-            SerializedFileWriter::new(file, Arc::new(parquet_schema), Default::default()).unwrap();
-
-        // Write a single row with a map containing one key-value pair.
-        // rep=0 for both columns: start of a new map.
-        // key def=2: map present (1) + key_value entry present (2), key is REQUIRED.
-        // value def=3: map present (1) + key_value entry present (2) + value present (3).
-        let mut row_group = writer.next_row_group().unwrap();
-        {
-            let mut col = row_group.next_column().unwrap().unwrap();
-            col.typed::<ByteArrayType>()
-                .write_batch(
-                    &[parquet::data_type::ByteArray::from("event_time")],
-                    Some(&[2]),
-                    Some(&[0]),
-                )
-                .unwrap();
-            col.close().unwrap();
-        }
-        {
-            let mut col = row_group.next_column().unwrap().unwrap();
-            col.typed::<Int96Type>()
-                .write_batch(&[int96_val], Some(&[3]), Some(&[0]))
-                .unwrap();
-            col.close().unwrap();
-        }
-        row_group.close().unwrap();
-        writer.close().unwrap();
-
-        let iceberg_schema = Arc::new(
-            Schema::builder()
-                .with_schema_id(1)
-                .with_fields(vec![
-                    NestedField::optional(
-                        1,
-                        "ts_map",
-                        Type::Map(crate::spec::MapType {
-                            key_field: NestedField::required(
-                                2,
-                                "key",
-                                Type::Primitive(PrimitiveType::String),
-                            )
-                            .into(),
-                            value_field: NestedField::optional(
-                                3,
-                                "value",
-                                Type::Primitive(PrimitiveType::Timestamp),
-                            )
-                            .into(),
-                        }),
-                    )
-                    .into(),
-                ])
-                .build()
-                .unwrap(),
-        );
-
-        let batches = read_int96_batches(&file_path, iceberg_schema, vec![1]).await;
-
-        assert_eq!(batches.len(), 1);
-        let map_array = batches[0]
-            .column(0)
-            .as_any()
-            .downcast_ref::<arrow_array::MapArray>()
-            .expect("Expected MapArray");
-        let ts_array = map_array
-            .values()
-            .as_any()
-            .downcast_ref::<TimestampMicrosecondArray>()
-            .expect("Expected TimestampMicrosecondArray as map values");
-
-        assert_eq!(
-            ts_array.value(0),
-            expected_micros,
-            "INT96 in map: got {}, expected {expected_micros}",
-            ts_array.value(0)
-        );
     }
 }
